refactor: reorganize currency and chain constants
- Moved currency and chain-related constants to dedicated files for better organization and maintainability.
- Updated imports across various components and schemas to reflect the new structure, ensuring consistent usage throughout the application.
