+++ src/lib/constants/currencies.ts
